Some "cast to pointer from integer of different size" warnings removed.
git-svn-id: https://svn.aros.org/svn/aros/trunk@54887 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/dopus/Config/main.c
@@ -1212,7 +1212,7 @@ int len,num;
     req.strbuf=buf;
     req.strlen=len;
     req.flags=num|SRF_BORDERS|SRF_RECESSHI|SRF_EXTEND;
-    req.value=(int)&stringex;
+    req.value=(long)&stringex;
     req.font=NULL;
     req.title="ConfigOpus";
     busy();
